+++ packages/fighting-design/image/src/interface.d.ts
@@ -3,3 +3,7 @@ export type imageFit = 'fill' | 'contain' | 'cover' | 'none' | 'scale-down' | ''
 export interface callbackInterface {
   (params: boolean, width: number): void
 }
+
+export interface handleCloseInterface {
+  (evt: MouseEvent): void
+}