+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - `f-image` 组件 `rootMargin` `width` `height` 属性支持 `number` 类型
 - `f-avatar` 组件 `rootMargin` 属性支持 `number` 类型
 - 新增 `f-image-preview` 组件
+- `f-image` 组件新增 `click` 回调
 
 **other**